Refactoring attributes and gpu data
