+++ lib/core/src/display/symbol.rs
@@ -19,9 +19,9 @@ use crate::debug::stats::Stats;
 use crate::display::render::webgl::Context;
 use crate::display::render::webgl;
 use crate::system::gpu::buffer::IsBuffer;
-use crate::system::gpu::scope::uniform::AnyUniform;
-use crate::system::gpu::scope::uniform::AnyUniformOps;
-use crate::system::gpu::scope::uniform::UniformScope;
+use crate::system::gpu::data::uniform::AnyUniform;
+use crate::system::gpu::data::uniform::AnyUniformOps;
+use crate::system::gpu::data::uniform::UniformScope;
 use crate::display::symbol::geometry::primitive::mesh;
 use crate::promote;
 use crate::promote_all;
