+++ lib/core/src/display/symbol/registry.rs
@@ -14,8 +14,8 @@ use crate::display::symbol;
 use crate::promote;
 use crate::promote_all;
 use crate::promote_symbol_types;
-use crate::system::gpu::scope::uniform::Uniform;
-use crate::system::gpu::scope::uniform::UniformScope;
+use crate::system::gpu::data::uniform::Uniform;
+use crate::system::gpu::data::uniform::UniformScope;
 use crate::system::web::group;
 use crate::system::web::Logger;
 use eval_tt::*;
