+++ lib/core/src/display/world.rs
@@ -27,8 +27,8 @@ use crate::display::shape::text::font::Fonts;
 use crate::debug::monitor;
 use crate::debug::monitor::Monitor;
 use crate::debug::monitor::Panel;
-use crate::system::gpu::scope::uniform::UniformScope;
-use crate::system::gpu::scope::uniform::Uniform;
+use crate::system::gpu::data::uniform::UniformScope;
+use crate::system::gpu::data::uniform::Uniform;
 use wasm_bindgen::prelude::Closure;
 use wasm_bindgen::{JsCast, JsValue};
