+++ lib/core/src/display/world/workspace.rs
@@ -22,7 +22,7 @@ use crate::system::web::group;
 use crate::system::web::Logger;
 use crate::system::web::resize_observer::ResizeObserver;
 use crate::system::web;
-use crate::system::gpu::scope::uniform::UniformScope;
+use crate::system::gpu::data::uniform::UniformScope;
 
 
 use eval_tt::*;
