+++ lib/core/src/system/gpu.rs
@@ -1,3 +1,2 @@
 pub mod buffer;
 pub mod data;
-pub mod scope;
