+++ lib/core/src/system/gpu/data/uniform.rs
