--- lib/core/src/system/gpu/scope.rs
@@ -1,18 +1,0 @@
-//! This module defines data types representing attributes and uniforms.
-//!   - Uniforms are per-primitive parameters (constant during an entire draw call).
-//!   - Attributes are per-vertex parameters (typically positions, normals, colors, UVs, ...).
-//!
-//! To learn more about these concepts, follow the link:
-//! https://www.khronos.org/opengl/wiki/Type_Qualifier_(GLSL)
-
-pub mod attribute;
-pub mod uniform;
-
-
-
-// =================
-// === Reexports ===
-// =================
-
-pub use attribute::*;
-pub use uniform::*;
